+++ commerce/interaction-api/src/main/java/ru/yandex/practicum/dto/order/TotalCostRequest.java
@@ -5,12 +5,14 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
 public class TotalCostRequest {
 
-    private Double productsCost;
-    private Double deliveryCost;
+    private BigDecimal productsCost;
+    private BigDecimal deliveryCost;
 }
